main/prototype/requirements.txt
astroid==2.4.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
df2gspread==1.0.4
et-xmlfile==1.1.0
google-api-python-client==1.6.7
google-auth==1.28.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
httplib2==0.19.1
idna==2.10
imageio==2.9.0
imutils==0.5.4
isort==5.6.4
joblib==1.0.1
lazy-object-proxy==1.4.3
matplotlib==3.4.0
mccabe==0.6.1
multipledispatch==0.6.0
mysql-connector-python==8.0.24
networkx==2.5
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.7
pandas==1.2.3
Pillow==8.1.2
protobuf==3.15.8
PTable==0.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.2
secure-smtplib==0.1.1
six==1.15.0
threadpoolctl==2.1.0
tifffile==2021.3.17
toml==0.10.2
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
xlrd==2.0.1
